sidevm: move the timer example
